Enhance documentation and refactor user preferences system
This commit significantly expands the documentation, including updates to `ARCHITECTURE.md`, `CLAUDE.md`, and the creation of new files like `FRONTEND_REVIEW_REPORT.md` and `SOFTWARE-DESIGN-STANDARDS.md`.

The `README.md` was revised for a comprehensive overview of the Central Command system.

In the codebase, the `IUserPreferencesService` interface was introduced to manage user preferences, along with several related classes for handling user settings and dashboard layouts.

Refactoring efforts included removing AutoMapper in favor of direct mapping in `AddIncidentCommentCommandHandler.cs`, changing property types to `Guid?` for better type safety, and enhancing the API response structure in `ApiResponse.cs`.

New services for metrics collection and health checks were added, and improvements were made to data access patterns and domain entity encapsulation.

+++ apps/api/CentralCommand.Api/Controllers/StatisticsController.cs
@@ -116,7 +116,7 @@ public class StatisticsController : ControllerBase
             AverageResponseTime = statistics.AverageResponseTime,
             AverageUptime = statistics.AverageUptime,
             AverageErrorRate = statistics.AverageErrorRate,
-            TotalRequests = statistics.TotalRequests,
+            TotalRequests = (int)statistics.TotalRequests,
             PeakResponseTime = statistics.AverageResponseTime * 1.5, // Simulated
             MinResponseTime = statistics.AverageResponseTime * 0.5, // Simulated
             P95ResponseTime = statistics.AverageResponseTime * 1.3, // Simulated
